test(item-divider): migrate basic
