+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-button-end-ios-ltr-Mobile-Chrome-linux.png
