+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-button-end-md-ltr-Mobile-Firefox-linux.png
