+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-button-end-md-rtl-Mobile-Firefox-linux.png
