+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-icon-start-ios-ltr-Mobile-Firefox-linux.png
