+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-icon-start-ios-rtl-Mobile-Chrome-linux.png
