+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-icon-start-md-ltr-Mobile-Chrome-linux.png
