+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-icon-start-md-rtl-Mobile-Chrome-linux.png
