+++ core/src/components/item-divider/test/basic/item-divider.e2e.ts-snapshots/item-divider-text-md-rtl-Mobile-Safari-linux.png
